use double instead of float

+++ svae/lds/global_optimization.py
@@ -6,7 +6,7 @@ from distributions import NormalInverseWishart, MatrixNormalInverseWishart, expo
 def initialize_global_lds_parameters(n, scale=1.0):
     device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
 
-    nu = torch.tensor([n + 1]).float()
+    nu = torch.tensor([n + 1])
     Phi = 2 * scale * (n + 1) * torch.eye(n)
     mu_0 = torch.zeros(n)
     kappa = torch.tensor([1 / (2 * scale * n)])
@@ -21,9 +21,9 @@ def initialize_global_lds_parameters(n, scale=1.0):
         torch.zeros_like(nu)
     ).standard_to_natural(nu, Phi, M, K)
 
-    dynamics_prior = tuple([d.to(device) for d in dynamics_prior])
+    dynamics_prior = tuple([d.to(device).double() for d in dynamics_prior])
 
-    return init_state_prior.to(device), dynamics_prior
+    return init_state_prior.to(device).double(), dynamics_prior
 
 
 def prior_kld_lds(eta_theta, eta_theta_prior):
